# server/src/openapi/__fixtures__/openapi.baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hatchery API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "description": "エラー応答（実装の実際の形に準拠）"
      },
      "Worker": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "personality": {
            "type": "string",
            "maxLength": 500
          },
          "verbosity": {
            "type": "string",
            "enum": [
              "concise",
              "standard",
              "detailed"
            ]
          },
          "imageUrl": {
            "type": "string",
            "maxLength": 500,
            "format": "uri"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "displayName"
        ],
        "description": "AI ワーカー（id / displayName / role / personality）"
      },
...
      "CommunityWorkerAssignments": {
        "type": "object",
        "properties": {
          "workers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                "displayName": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 50
                }
              },
              "required": [
                "id",
                "displayName"
              ]
            }
          }
        },
        "required": [
          "workers"
        ],
        "description": "コミュニティ所属ワーカー一覧（id・displayName・#1079）"
      },
      "SetCommunityWorkers": {
        "type": "object",
        "properties": {
          "workerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "maxItems": 200
          }
        },
        "required": [
          "workerIds"
        ],
        "description": "コミュニティの所属ワーカーを置き換えるリクエストボディ（#1079）"
      },
...
    },
...
  },
  "paths": {
...
    "/api/admin/communities/{id}/workers": {
      "get": {
        "summary": "コミュニティの所属ワーカー一覧を取得（認証必須・admin のみ・#1079）",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "所属ワーカー一覧（id・displayName）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityWorkerAssignments"
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "admin 権限なし",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "コミュニティが存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "コミュニティの所属ワーカーを置き換える（認証必須・admin のみ・#1079）",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCommunityWorkers"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "置換後の所属ワーカー一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityWorkerAssignments"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー / 存在しない workerId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "admin 権限なし",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "コミュニティが存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/communities/{slug}/workers": {
      "get": {
        "summary": "community 所属の全ワーカー一覧を取得（認証不要・id 昇順・カーソルページネーション）",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 512,
              "description": "カーソル（直前ページ末尾ワーカー id の base64 エンコード）"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "1 ページあたりの取得件数（デフォルト 20・最大 100）"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "community 所属ワーカー一覧（id 昇順・カーソルページネーション）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Worker"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "maxLength": 512,
                      "description": "次ページ取得用カーソル。null の場合は末尾"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "コミュニティが存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}